[main] Refactored useNotification (was attempting to send notifications for notifications not coming from the database)

+++ src/hooks/useNotifications.ts
@@ -28,7 +28,7 @@ export const useNotifications = (): void => {
         const unsubscribe = toast.onChange(async (payload: ToastItem) => {
             const { data, status } = payload;
             const { notificationId } = data as ToastNotificationData;
-            if (status === "added") {
+            if (status === "added" && notificationId !== undefined) {
                 await NotificationService.removeNotification(notificationId);
             }
         });
